+++ tests/test_yaralyze.py
@@ -1,9 +1,12 @@
 """
 Tests for invoking yaralyze script from shell.
 """
+from argparse import ArgumentError
 from math import isclose
 from os import environ
-from subprocess import check_output
+from subprocess import CalledProcessError, check_output
+
+import pytest
 
 from yaralyzer.config import YARALYZE
 from yaralyzer.helpers.rich_text_helper import console
@@ -18,11 +21,18 @@ def test_help_option():
     assert len(help_text.split('\n')) > 50
 
 
+def test_argument_exceptions(il_tulipano_path, tulips_yara_path):
+    with pytest.raises(CalledProcessError):
+        _run_with_args(il_tulipano_path, '-Y', tulips_yara_path, '-re', 'tulip')
+
+
 def test_yaralyze(il_tulipano_path, tulips_yara_path, tulips_yara_regex):
     # yaralyze -y tests/file_fixtures/tulips.yara tests/file_fixtures/il_tulipano_nero.txt
     with_yara_file_output = _run_with_args(il_tulipano_path, '-Y', tulips_yara_path)
     # yaralyze -r 'tulip.{1,2500}tulip' tests/file_fixtures/il_tulipano_nero.txt
     with_pattern_output = _run_with_args(il_tulipano_path, '-re', tulips_yara_regex)
+    # yaralyze -dir tests/file_fixtures/ tests/file_fixtures/il_tulipano_nero.txt
+
     assert line_count(with_pattern_output) == line_count(with_yara_file_output)
     _assert_line_count_within_range(814, with_yara_file_output)