+++ tests/test_yaralyzer.py
@@ -19,7 +19,7 @@ def test_yaralyzer_with_patterns(il_tulipano_path, tulips_yara_regex):
 
 
 def test_yaralyzer_for_rules_dir(il_tulipano_path, tulips_yara_regex):
-    _test_yaralyze(Yaralyzer.for_rules_dir(dirname(il_tulipano_path), il_tulipano_path))
+    _test_yaralyze(Yaralyzer.for_rules_dirs([dirname(il_tulipano_path)], il_tulipano_path))
 
 
 def _test_yaralyze(yaralzyer: Yaralyzer) -> None: