+++ yaralyzer/__init__.py
@@ -24,6 +24,8 @@ def yaralyze():
         yaralyzer = Yaralyzer.for_rules_files(args.yara_rules_files, args.file_to_scan_path)
     elif args.yara_patterns:
         yaralyzer = Yaralyzer.for_patterns(args.yara_patterns, args.file_to_scan_path)
+    elif args.yara_rules_dirs:
+        yaralyzer = Yaralyzer.for_rules_dirs(args.yara_rules_dirs, args.file_to_scan_path)
     else:
         raise RuntimeError("No pattern or YARA file to scan against.")