+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Next Release
+* Add `--rules-dir` option for loading all `.yara` files in a directory
+* Add `--regex_modifier` option
+
 # 0.2.0
 * Add `Yaralyzer.for_rules_dir()` constructor to load all `.yara` files in a directory
 * Change command line arguments `-y` to `-Y` and `-r` to `-re`